# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zangoh AI Agent Supervisor API",
    "version": "1.0.0",
    "description": "REST API for the supervisor workstation. Identity is taken from the optional `x-supervisor-id` header (default `supervisor-001`). Real-time updates are pushed over WebSocket on the same port (see README)."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Toggle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "AgentConfig": {
        "type": "object",
        "properties": {
          "parameters": {
            "type": "object",
            "properties": {
              "temperature": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              "max_tokens": {
                "type": "integer",
                "minimum": 16,
                "maximum": 4096
              },
              "top_p": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              }
            }
          },
          "capabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Toggle"
            }
          },
          "knowledgeBases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Toggle"
            }
          },
          "escalationThresholds": {
            "type": "object",
            "properties": {
              "lowConfidence": {
                "type": "number"
              },
              "negativeSentiment": {
                "type": "number"
              },
              "responseTime": {
                "type": "number",
                "description": "seconds"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "maintenance"
            ]
          }
        }
      },
      "Agent": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "model": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "metrics": {
                "type": "object"
              }
            }
          },
          {
            "$ref": "#/components/schemas/AgentConfig"
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/intervene": {
      "post": {
        "tags": [
          "Intervention"
        ],
        "summary": "Take over a conversation from the AI (409 if another supervisor holds it)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "supervisorId": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                },
                "required": [
                  "conversationId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "intervention": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Held by another supervisor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/intervene/release": {
      "post": {
        "tags": [
          "Intervention"
        ],
        "summary": "Return control to the AI with guidance notes",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "supervisorNotes": {
                    "type": "string"
                  }
                },
                "required": [
                  "conversationId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "conversation": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agents/{id}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}/config": {
      "patch": {
        "tags": [
          "Agents"
        ],
        "summary": "Partially update agent configuration (validated)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "agent": {
                      "$ref": "#/components/schemas/Agent"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analytics/overview": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Headline metrics + daily trends",
        "parameters": [
          {
            "name": "timeRange",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/agents": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Per-agent stats + trends",
        "parameters": [
          {
            "name": "timeRange",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ]
            }
          },
          {
            "name": "agentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/issues": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Top issues from conversation tags",
        "parameters": [
          {
            "name": "timeRange",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  }
}